requirements.txt
python-oracledb
cryptography